requirements.txt
faiss-cpu             
ffmpeg                
gradio                
gTTS                  
langchain-community   
langchain-groq        
langchain-huggingface 
load-dotenv           
PyAudio               
pygame                
pymupdf4llm           
SpeechRecognition    